Source link sorting by angle
Incoming source links to a node are sorted by the angle between the
source and target nodes, rather than their difference in y. It produces
cleaner diagrams, and has a customizable weighting factor to give more
weight to the x difference between nodes. The default value of 3 seems
to work well.